requirements.txt
djangorestframework-jsonapi==4.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
Django==3.0.5
django-cors-headers==3.6.0
django-extensions==3.1.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
geoip2==4.1.0
geos==0.2.2
postgis==1.0.4
proj==0.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.25.1
aiohttp==3.7.3